refactor(core): move the swipe-gesture machinery to binding/swipe.rs
detect_swipe and SwipeAccumulator are input-processing runtime, not
part of the Action vocabulary the rest of binding.rs defines.
